feat: remove redunctunt code v2
