+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+84",
+  "version": "0.9.0+85",
   "private": true,
   "type": "module",
   "scripts": {
